+++ src/main/java/com/parksupark/soomjae/server/auth/common/jwt/DefaultJwtHandler.java
@@ -1,43 +1,23 @@
 package com.parksupark.soomjae.server.auth.common.jwt;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.util.Base64;
 import java.util.Date;
 
 @Component
-public class DefaultJwtHandler implements JwtHandler {
-    private final SecretKey key;
+public class DefaultJwtHandler extends AbstractJwtHandler {
 
     public DefaultJwtHandler(@Value("${jwt.secret}") String secret) {
-        byte[] keyBytes = Base64.getDecoder().decode(secret);
-        this.key = new SecretKeySpec(keyBytes, "HmacSHA256");
+        super(secret);
     }
 
     @Override
-    public String generate(String subject) {
-        long nowMillis = System.currentTimeMillis();
-        long expirationMillis = nowMillis + 1000 * 60 * 60;
-
-        return Jwts.builder()
-                .subject(subject)
-                .signWith(key)
-                .issuedAt(new Date(nowMillis))
-                .expiration(new Date(expirationMillis))
-                .compact();
+    protected Date getIssuedAt() {
+        return new Date();
     }
 
     @Override
-    public Claims parse(String token) {
-        return Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload();
+    protected Date getExpiration() {
+        return new Date(System.currentTimeMillis() + 3600000); // 1시간 후
     }
 }
